+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Feat
+
+- **agent**: add optional `onPollingStarted` callback when making update calls (#1345)
+
 ## v5.3.1 (2026-04-23)
 
 ### Refactor